% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_test.R
\name{struct_test}
\alias{struct_test}
\alias{struct_test.ergmito}
\alias{struct_test.formula}
\title{Semiparametric test of structural dependence}
\usage{
struct_test(x, y, stat, R, ...)

\method{struct_test}{ergmito}(x, y, stat, R, ...)

\method{struct_test}{formula}(x, y, stat, R, ergmito.args = list(), ...)
}
\arguments{
\item{x}{Either a model, or an \link[ergmito:ergmito]{ergmito::ergmito} fit.}

\item{y}{A numeric vector of length \code{length(x)}}

\item{stat}{Function that receives two arguments, a network and a scalar.}

\item{R}{Integer. Number of replicates.}

\item{...}{(Ignored)}

\item{ergmito.args}{A list with arguments passed to \link[ergmito:ergmito]{ergmito::ergmito}.}
}
\description{
Semiparametric test of structural dependence
}
